# Test, benchmark and optimize the input system