Adds full test coverage to key and input related systems. The tests are generic and might be missing some edge cases but those should be added as they pop up.

Adds benchmarks to prevent performance degradation with future changes. Again generic tests but they did point out some issues that are now resolved.

Tweaked the key manager to be faster in almost every use case:

- Removed a lot of branching behavior
- Removed the use of 'defer' due to the high cost

---

Over all speed up from original code is 3x on all benchmarks below latest result on my system:

```
 nitya > ⋯ > src > engo.io > engo > go test -bench=. -benchmem
PASS
BenchmarkInputMgr_AxisCleanState-4       2000000               760 ns/op               0 B/op          0 allocs/op
BenchmarkInputMgr_AxisFilledState-4      2000000               896 ns/op               0 B/op          0 allocs/op
BenchmarkInputMgr_ButtonCleanState-4     1000000              1071 ns/op               0 B/op          0 allocs/op
BenchmarkInputMgr_ButtonFilledState-4    1000000              1217 ns/op               0 B/op          0 allocs/op
BenchmarkKeyMgr_CleanState-4             5000000               327 ns/op               0 B/op          0 allocs/op
BenchmarkKeyMgr_FilledState-4            5000000               395 ns/op               0 B/op          0 allocs/op
BenchmarkKeyMgr_CleanUpdate-4            5000000               371 ns/op               0 B/op          0 allocs/op
BenchmarkKeyMgr_FilledUpdate-4           3000000               426 ns/op               0 B/op          0 allocs/op
BenchmarkKeyMgr_CleanSubOptimal-4        1000000              1104 ns/op               0 B/op          0 allocs/op
BenchmarkKeyMgr_FilledSubOptimal-4       1000000              1308 ns/op               0 B/op          0 allocs/op
ok      engo.io/engo    18.008s
```

No allocs means no garbage generated and that matters as well ;)

---

Edit: More changed and with tests in place this is now up for merge again. Going any further would need changes that belong in a separated commit.