@nitya-dev 2016-05-27
Maybe hold off on merge. I'm reworking the whole input mechanics for my own project, might want to merge that in staid if it is ready. Based of the current system and my imput manager in unity.

Key Changes:
* Key and mouse combo actions
* Mouse buttons behave like keys
* Events when actions change state
* Store state for registered actions once

'Actions' would be axis, button and sequence objects that are registered. Stored state would update all registered actions once so no complex methods need to be called when checking the state of an action inside update functions.