@nitya-dev 2016-06-03
Updated with test and benchmark. Initial changes gave +-10% speed increase, the last commit just see ...

Before:
```
$ nitya > ⋯ > src > engo.io > engo > go test -bench=.
PASS
BenchmarkKeyMgrCleanState-4              1000000              1102 ns/op
BenchmarkKeyMgrCleanUpdate-4             1000000              1149 ns/op
BenchmarkKeyMgrFilledState-4             1000000              1287 ns/op
BenchmarkKeyMgrFilledUpdate-4             300000              4082 ns/op
BenchmarkKeyMgrCleanStateSubOptimal-4     300000              4201 ns/op
ok      engo.io/engo    6.156s
```

After:
```
$ nitya > ⋯ > src > engo.io > engo > go test -bench=.
PASS
BenchmarkKeyMgrCleanState-4              5000000               318 ns/op
BenchmarkKeyMgrCleanUpdate-4             5000000               362 ns/op
BenchmarkKeyMgrFilledState-4             5000000               379 ns/op
BenchmarkKeyMgrFilledUpdate-4             500000              3129 ns/op
BenchmarkKeyMgrCleanStateSubOptimal-4    1000000              1121 ns/op
ok      engo.io/engo    9.114s
```

Feel free to post your results and thoughts, always interested to see numbers from other systems.