@otraore 2016-06-03
Thank you @nitya-dev . I didn't realize about the cost of defer until now (https://github.com/nitya-dev/engo/commit/6da5bb9a10a7cfd5a76ca294c67fc49c171c41ee#diff-3afb33cd4bbb018c7856457d229b4586L41 and talks with @hydroflame ).